2nd round of refactor
